fix(salary-type): unify salary type throughout the application

+++ booking-service/src/main/java/com/github/bovvver/bookingmanagement/bookingdraftcreation/BookOfferRequest.java
@@ -3,8 +3,9 @@ package com.github.bovvver.bookingmanagement.bookingdraftcreation;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record BookOfferRequest(@NotNull UUID offerId,
-                               @PositiveOrZero Double salary) {
+                               @PositiveOrZero BigDecimal salary) {
 }
